requirements.txt
Flask
Pillow
torch 
torchvision
numpy
ultralytics 
matplotlib 
opencv-python 
PyYAML 
